Simple Fog of War with Line of Sight.
Applied a Layer name to all the Enemies and Players to "Enemy" & "Player".
added a script to the enemy AI
added a script to the players
added a ball of light to the players.
added a shader to the canvas for a fog of war effect.

Enemies no long render when out of the line of sight from players.
Note: This probably won't work if it's a 1v1 but when its 1vs AI its fine.

To Do:
Add this light to occupied buildings.